fixed Box dtype warning

+++ osim/env/utils/mygym.py
@@ -7,7 +7,7 @@ class Specification:
         self.timestep_limit = timestep_limit
 
 def convert_to_gym(space):
-    return spaces.Box(np.array(space[0]), np.array(space[1]) )
+    return spaces.Box(np.array(space[0]), np.array(space[1]), dtype=np.float32 )
 
 def gymify_env(env):
     env.action_space = convert_to_gym(env.action_space)
